refactor: split reorder.rs into module directory with tests in sibling file
The reorder.rs file had grown to ~62 KB with embedded test modules accounting
for roughly two-thirds of the line count. Converting to a `reorder/` directory
module mirrors the established `work_queue/` layout: `mod.rs` for
implementation, `tests.rs` for the test modules. No public API or test
behaviour changes - tests retain their original names, assertions, and
groupings under inner `tests`, `adaptive_tests`, and `passthrough_tests`
modules.
